ARGO/requirements.txt
# ARGO项目依赖包
# ARGO - Adaptive Retrieval with Guided Optimization

# 核心依赖
numpy>=1.21.0
scipy>=1.7.0
gym>=0.21.0

# RAG相关
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
faiss-cpu>=1.7.4  # 向量检索库(CPU版本)
# faiss-gpu>=1.7.4  # GPU版本(如果有GPU)
accelerate>=0.26.0
bitsandbytes>=0.43.0
chromadb>=0.5.5
pypdf>=4.3.1
python-docx>=0.8.11

# 文本处理
nltk>=3.8
spacy>=3.5.0

# 数据处理
pandas>=1.5.0
scikit-learn>=1.2.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.12.0

# 工具
tqdm>=4.65.0
pyyaml>=6.0

# 可选：LLM集成
# openai>=1.0.0
# langchain>=0.1.0
